chore(utils): moving generateRandomCoordinates to a utility file
